requirements.txt
torch>=2.0.0
torchvision
numpy
pandas
tqdm
matplotlib
scikit-learn
pillow
scipy
